Add smooth movement interpolation and combat initiation systems 10:27pm 7/17/2025
Introduces InterpolationComponent and InterpolationSystem for smooth entity movement, refactors AI and player movement to use visual interpolation, and adds CombatInitiationSystem to safely handle combat start requests. Updates AI logic to use pathfinding and time budgets, modifies rendering to display movement and intent indicators with new sprites, and adjusts time passage/event handling to support fractional seconds. Refactors related systems and components for consistency and improved gameplay flow.

+++ ECS/AIComponent.cs
@@ -19,7 +19,7 @@
         /// The amount of in-game time (in seconds) the AI has to spend on its actions.
         /// This is granted when the player performs an action.
         /// </summary>
-        public int ActionTimeBudget { get; set; } = 0;
+        public float ActionTimeBudget { get; set; } = 0;
 
         public IComponent Clone()
         {
